+++ lib/main.js
@@ -1,9 +1,9 @@
 'use strict';
-const process = require('./process-adapter');
+const worker = require('./test-worker');
+const adapter = require('./process-adapter');
 const serializeError = require('./serialize-error');
 const globals = require('./globals');
 const Runner = require('./runner');
-const send = process.send;
 
 const opts = globals.options;
 const runner = new Runner({
@@ -13,7 +13,7 @@ const runner = new Runner({
 });
 
 // Note that test files have require('ava')
-require('./test-worker').avaRequired = true;
+worker.avaRequired = true;
 
 // If fail-fast is enabled, use this variable to detect
 // that no more tests should be logged
@@ -39,7 +39,7 @@ function test(props) {
 		props.error = null;
 	}
 
-	send('test', props);
+	adapter.send('test', props);
 
 	if (hasError && opts.failFast) {
 		isFailed = true;
@@ -50,7 +50,7 @@ function test(props) {
 function exit() {
 	const stats = runner._buildStats();
 
-	send('results', {stats});
+	adapter.send('results', {stats});
 }
 
 globals.setImmediate(() => {
@@ -58,11 +58,11 @@ globals.setImmediate(() => {
 	const numberOfTests = runner.tests.testCount;
 
 	if (numberOfTests === 0) {
-		send('no-tests', {avaRequired: true});
+		adapter.send('no-tests', {avaRequired: true});
 		return;
 	}
 
-	send('stats', {
+	adapter.send('stats', {
 		testCount: numberOfTests,
 		hasExclusive
 	});
