+++ lib/process-adapter.js
@@ -1,23 +1,18 @@
 'use strict';
 const fs = require('fs');
 const path = require('path');
-const chalk = require('chalk');
+const debug = require('debug')('ava');
 const sourceMapSupport = require('source-map-support');
 const installPrecompiler = require('require-precompiled');
 
-const debug = require('debug')('ava');
+// Parse and re-emit AVA messages
+process.on('message', message => {
+	if (!message.ava) {
+		return;
+	}
 
-// Check if the test is being run without AVA cli
-const isForked = typeof process.send === 'function';
-
-if (!isForked) {
-	const fp = path.relative('.', process.argv[1]);
-
-	console.log();
-	console.error('Test files must be run with the AVA CLI:\n\n    ' + chalk.grey.dim('$') + ' ' + chalk.cyan('ava ' + fp) + '\n');
-
-	process.exit(1); // eslint-disable-line unicorn/no-process-exit
-}
+	process.emit(message.name, message.data);
+});
 
 exports.send = (name, data) => {
 	process.send({
@@ -26,11 +21,6 @@ exports.send = (name, data) => {
 		ava: true
 	});
 };
-
-exports.on = process.on.bind(process);
-exports.emit = process.emit.bind(process);
-exports.exit = process.exit.bind(process);
-exports.env = process.env;
 
 const opts = JSON.parse(process.argv[2]);
 exports.opts = opts;
