+++ lib/test-worker.js
@@ -1,37 +1,55 @@
 'use strict';
-/* eslint-disable import/order */
+
+// Check if the test is being run without AVA cli
+{
+	/* eslint-disable import/order */
+	const path = require('path');
+	const chalk = require('chalk');
+
+	const isForked = typeof process.send === 'function';
+	if (!isForked) {
+		const fp = path.relative('.', process.argv[1]);
+
+		console.log();
+		console.error('Test files must be run with the AVA CLI:\n\n    ' + chalk.grey.dim('$') + ' ' + chalk.cyan('ava ' + fp) + '\n');
+
+		process.exit(1); // eslint-disable-line unicorn/no-process-exit
+	}
+}
+
+/* eslint-enable import/order */
+const Bluebird = require('bluebird');
+const currentlyUnhandled = require('currently-unhandled')();
 const isObj = require('is-obj');
-const process = require('./process-adapter');
-
-const opts = process.opts;
-const testPath = opts.file;
-
-// Bind globals first before anything has a chance to interfere
+const adapter = require('./process-adapter');
 const globals = require('./globals');
+const serializeError = require('./serialize-error');
+const throwsHelper = require('./throws-helper');
+
+const opts = adapter.opts;
+const testPath = opts.file;
 globals.options = opts;
-const Promise = require('bluebird');
 
 // Bluebird specific
-Promise.longStackTraces();
+Bluebird.longStackTraces();
 
 (opts.require || []).forEach(require);
 
-process.installSourceMapSupport();
+adapter.installSourceMapSupport();
+adapter.installPrecompilerHook();
 
-const currentlyUnhandled = require('currently-unhandled')();
-const serializeError = require('./serialize-error');
-const send = process.send;
-const throwsHelper = require('./throws-helper');
+const dependencies = [];
+adapter.installDependencyTracking(dependencies, testPath);
 
 // Check if test files required ava and show error, when they didn't
 exports.avaRequired = false;
 
-process.installPrecompilerHook();
-
-const dependencies = [];
-process.installDependencyTracking(dependencies, testPath);
-
 require(testPath); // eslint-disable-line import/no-dynamic-require
+
+// If AVA was not required, show an error
+if (!exports.avaRequired) {
+	adapter.send('no-tests', {avaRequired: false});
+}
 
 process.on('unhandledRejection', throwsHelper);
 
@@ -51,30 +69,7 @@ process.on('uncaughtException', exception => {
 			stack: err.stack
 		};
 	}
-	send('uncaughtException', {exception: serialized});
-});
-
-// If AVA was not required, show an error
-if (!exports.avaRequired) {
-	send('no-tests', {avaRequired: false});
-}
-
-// Parse and re-emit AVA messages
-process.on('message', message => {
-	if (!message.ava) {
-		return;
-	}
-
-	process.emit(message.name, message.data);
-});
-
-process.on('ava-exit', () => {
-	// Use a little delay when running on AppVeyor (because it's shit)
-	const delay = process.env.AVA_APPVEYOR ? 100 : 0;
-
-	globals.setTimeout(() => {
-		process.exit(0); // eslint-disable-line xo/no-process-exit
-	}, delay);
+	adapter.send('uncaughtException', {exception: serialized});
 });
 
 let tearingDown = false;
@@ -87,28 +82,26 @@ process.on('ava-teardown', () => {
 
 	let rejections = currentlyUnhandled();
 
-	if (rejections.length === 0) {
-		exit();
-		return;
+	if (rejections.length > 0) {
+		rejections = rejections.map(rejection => {
+			let reason = rejection.reason;
+			if (!isObj(reason) || typeof reason.message !== 'string') {
+				reason = {
+					message: String(reason)
+				};
+			}
+			return serializeError(reason);
+		});
+
+		adapter.send('unhandledRejections', {rejections});
 	}
 
-	rejections = rejections.map(rejection => {
-		let reason = rejection.reason;
-		if (!isObj(reason) || typeof reason.message !== 'string') {
-			reason = {
-				message: String(reason)
-			};
-		}
-		return serializeError(reason);
-	});
-
-	send('unhandledRejections', {rejections});
-	globals.setTimeout(exit, 100);
-});
-
-function exit() {
 	// Include dependencies in the final teardown message. This ensures the full
 	// set of dependencies is included no matter how the process exits, unless
 	// it flat out crashes.
-	send('teardown', {dependencies});
-}
+	adapter.send('teardown', {dependencies});
+});
+
+process.on('ava-exit', () => {
+	process.exit(0); // eslint-disable-line xo/no-process-exit
+});
